add static data add rule system init_data function add

+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_item_id_table.hpp
@@ -6,10 +6,10 @@
 class critemid
 {
   public:
-    uint8_t i_id;
+    uint64_t i_id;
     uint8_t i_type;
 public:
-    uint8_t primary_key() const {return i_id;}
+    uint64_t primary_key() const {return i_id;}
 };
 
 typedef multi_index<N(critemid),critemid> item_id_data_table;
