+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_item_tier_table.hpp
@@ -6,10 +6,10 @@
 class critemtier
 {
   public:
-    uint8_t i_tier;
+    uint64_t i_tier;
     uint8_t i_level;
 public:
-    uint8_t primary_key() const {return i_tier;}
+    uint64_t primary_key() const {return i_tier;}
 };
 
 typedef multi_index<N(critemtier),critemtier> item_tier_data_table;
