+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_monster_id_table.hpp
@@ -6,9 +6,9 @@
 class crmid
 {
   public:
-    uint8_t m_id;
+    uint64_t m_id;
 public:
-    uint8_t primary_key() const {return m_id;}
+    uint64_t primary_key() const {return m_id;}
 };
 
 typedef multi_index<N(crmid),crmid> monster_id_data_table;
